Add option to skip vanilla LDC benchmark

+++ automate_benchmark.py
@@ -168,6 +168,7 @@ if __name__ == '__main__':
     parser.add_argument('-b', '--serial-baudrate', type=int, default=115200, help='')
     parser.add_argument('--platformio_path', default='pio', help="path to the platformio executable. default to 'pio'.")
     parser.add_argument('-v', '--validate-p', action='store_true', help='')
+    parser.add_argument('-n', '--skip-vanilla', action='store_true', help='skip running benchmark on vanilla LDC (for board with limited flash)')
     parser.add_argument('-t', '--num-test-sample', type=int, default=5, help='maximum value of the input')
     args = parser.parse_args()
 
@@ -191,22 +192,23 @@ if __name__ == '__main__':
     raw_predict = np.load(f'test_data/{args.model}_{args.df}s_raw_predict.npy')
 
     # test vanilla LDC
-    print ('Testing vanilla LDC...')
-    os.symlink(f'../../../model/model_{args.model}_{args.df}sb.h', f'{project_dir}/{get_include_dir_name(project_dir)}/model.h')
-    os.symlink(f'../../../model/model_{args.model}_{args.df}sb.c', f'{project_dir}/{get_source_dir_name(project_dir)}/model.c')
-    if args.word_size == 8 or args.word_size == 16:
-        configs = vanilla_ldc_configs_8_16bit
-    else:
-        configs = vanilla_ldc_configs_32_bit
-    results = []
-    for i, build_flag in enumerate(configs):
-        print (f'    Configuration {i+1}/{len(configs)}...')
-        ram, flash, avg_runtime = upload_and_benchmark(project_dir, build_flag + common_build_flag, sample, predict, raw_predict, args)
-        results.append([ram, flash, avg_runtime])
-    best_result = min(results, key=lambda e: e[2])
-    table.add_row(['Vanilla LDC', best_result[0], best_result[1], best_result[2]])
-    os.remove(f'{project_dir}/{get_include_dir_name(project_dir)}/model.h')
-    os.remove(f'{project_dir}/{get_source_dir_name(project_dir)}/model.c')
+    if not args.skip_vanilla:
+        print ('Testing vanilla LDC...')
+        os.symlink(f'../../../model/model_{args.model}_{args.df}sb.h', f'{project_dir}/{get_include_dir_name(project_dir)}/model.h')
+        os.symlink(f'../../../model/model_{args.model}_{args.df}sb.c', f'{project_dir}/{get_source_dir_name(project_dir)}/model.c')
+        if args.word_size == 8 or args.word_size == 16:
+            configs = vanilla_ldc_configs_8_16bit
+        else:
+            configs = vanilla_ldc_configs_32_bit
+        results = []
+        for i, build_flag in enumerate(configs):
+            print (f'    Configuration {i+1}/{len(configs)}...')
+            ram, flash, avg_runtime = upload_and_benchmark(project_dir, build_flag + common_build_flag, sample, predict, raw_predict, args)
+            results.append([ram, flash, avg_runtime])
+        best_result = min(results, key=lambda e: e[2])
+        table.add_row(['Vanilla LDC', best_result[0], best_result[1], best_result[2]])
+        os.remove(f'{project_dir}/{get_include_dir_name(project_dir)}/model.h')
+        os.remove(f'{project_dir}/{get_source_dir_name(project_dir)}/model.c')
 
     # test binary LDC
     print ('Testing binary LDC...')
